+++ lib/crass.rb
@@ -11,4 +11,12 @@ module Crass
     Parser.parse_stylesheet(input, options)
   end
 
+  # Parses _input_ as a string of CSS properties (such as the contents of an
+  # HTML element's `style` attribute) and returns a parse tree.
+  #
+  # See {Tokenizer#initialize} for _options_.
+  def self.parse_properties(input, options = {})
+    Parser.parse_properties(input, options)
+  end
+
 end